+++ task-parser/src/main/java/com/oppo/cloud/parser/service/writer/MysqlWriter.java
@@ -67,7 +67,7 @@ public class MysqlWriter {
         try {
             String sql = "UPDATE bdmp_cluster.t_script_sql_diagnose_result SET score=?,score_content=?,diagnose_result=?,data_date=? where script_name =?";
             ps = connection.prepareStatement(sql);
-            ps.setInt(1, sqlScoreAbnormal.getScore());
+            ps.setDouble(1, sqlScoreAbnormal.getScore());
             ps.setString(2, sqlScoreAbnormal.getScoreContent());
             ps.setString(3, sqlScoreAbnormal.getDiagnoseResult());
             ps.setString(4, String.valueOf(System.currentTimeMillis()));
